test: Test Region
